+++ lib/kb_gtdbtk/core/api_translation.py
@@ -19,6 +19,9 @@ class GTDBTKParams(_NamedTuple):
     min_perc_aa: float
     ''' The mimimum sequence alignment in percent. '''
 
+    overwrite_tax: int
+    ''' Boolean to decode wheter to overwrite an existing Taxonomy field in input Genome. '''
+
 
 def get_gtdbtk_params(input_params: Dict[str, object]) -> GTDBTKParams:
     '''
@@ -51,4 +54,10 @@ def get_gtdbtk_params(input_params: Dict[str, object]) -> GTDBTKParams:
     if type(wsid) != int or _cast(int, wsid) < 1:
         raise ValueError('workspace_id is required and must be an integer > 0')
 
-    return GTDBTKParams(_cast(str, ref), _cast(int, wsid), _cast(float, min_perc_aa) * 1.0)
+    overwrite_tax = input_params.get('overwrite_tax', 1)
+    if type(overwrite_tax) != int \
+       or (_cast(int, overwrite_tax) != 0 and _cast(int, overwrite_tax) != 1):
+        raise ValueError('overwrite_tax is required and must be an integer [0,1]')
+
+    
+    return GTDBTKParams(_cast(str, ref), _cast(int, wsid), _cast(float, min_perc_aa) * 1.0, _cast(int, overwrite_tax))